refactor(darkmatter): simplify style parsing with guard clauses and fix width comparison
- use guard clause pattern for deprecated alignment check in scan_deprecated_enum_aliases
- fix truncate_to_cells width comparison to not redundantly cast usize to usize

+++ darkmatter/lib/src/style/bespoke.rs
@@ -495,7 +495,7 @@ fn truncate_to_cells(text: &str, max_cells: u16) -> String {
         acc.push(ch);
         used += cw;
     }
-    if reserve_ellipsis && UnicodeWidthStr::width(text) as usize > max {
+    if reserve_ellipsis && UnicodeWidthStr::width(text) > max {
         acc.push('…');
     }
     acc
